Block reward table fix
