+++ .env.example.kadena
@@ -52,3 +52,4 @@ NEXT_PUBLIC_WALLET_CONNECT_PROJECT_ID=123456
 NEXT_PUBLIC_NETWORK_RPC_URL=https://evm-testnet.chainweb.com/chainweb/0.0/evm-testnet/${CHAINWEB_CHAIN_ID}/evm/rpc/
 
 NEXT_PUBLIC_OG_IMAGE_URL=public/static/kadena_evm_og_image.png
+NEXT_PUBLIC_VIEWS_BLOCK_HIDDEN_FIELDS=[\"total_reward\"]
